Tested RP2040-ESP communication library, works with UF2 V1.20. Errors occur in the second thread with UF2 V1.18

+++ RP2040-FRW/src/peripherals.py
@@ -50,7 +50,7 @@ class LightSensor:
         return(self.adc.read_u16())
 
     def getLightVoltage(self):
-        return(self._voltage*self.adc.read_u16()/65535)
+        return int(self._voltage*self.adc.read_u16()/65535)
     
     def getLightPercentage(self):
         raw_value = self.adc.read_u16()
